Avoid re-run to save run time

+++ scripts/WebDownloader.py
@@ -1,6 +1,6 @@
 import os
-import TarExtractor
 import subprocess as sp
+import TarExtractor
 import ColorTextWriter
 
 class WebDownloader:
@@ -18,19 +18,24 @@ class WebDownloader:
 
         ctw = ColorTextWriter.ColorTextWriter()
 
-        print(ctw.CRED + 'Downloading ' + ctw.CBLUE + self.download_file + ctw.CRED + ' ...' + ctw.CEND + '\n')
+        if len(os.listdir(outdir + '/')) == 0:
 
-        sp.call('wget ' + self.input_path + self.download_file, shell=True)
+            print(ctw.CRED + 'Downloading ' + ctw.CBLUE + self.download_file + ctw.CRED + ' ...' + ctw.CEND + '\n')
 
-        print(ctw.CBLUE + self.download_file + ctw.CRED + ' Downloaded!!!' + ctw.CEND)
+            sp.call('wget ' + self.input_path + self.download_file, shell=True)
 
-        dir_path = os.path.dirname(os.path.realpath(__file__))
-        te = TarExtractor.TarExtractor(dir_path + '/' + self.download_file, self.output_dir)
-        te.tar_extractor()
+            print(ctw.CBLUE + self.download_file + ctw.CRED + ' Downloaded!!!' + ctw.CEND)
 
-        if self.download_file.endswith('.tar.gz'):
-            sp.call('rm -r ' + self.download_file, shell=True)
+            dir_path = os.path.dirname(os.path.realpath(__file__))
+            te = TarExtractor.TarExtractor(dir_path + '/' + self.download_file, self.output_dir)
+            te.tar_extractor()
+
+            if self.download_file.endswith('.tar.gz'):
+                sp.call('rm -r ' + self.download_file, shell=True)
+
+            else:
+                new_file = self.download_file.split('.gz')[0]
+                sp.call(['mv', new_file, outdir + '/' + new_file])
 
         else:
-            new_file = self.download_file.split('.gz')[0]
-            sp.call(['mv', new_file, outdir + '/' + new_file])
+            print(ctw.CRED + "Destination directory contain files. Ignore download!!!" + ctw.CEND + '\n')
